Refactoring to increase code reuse.
